Rename `rustc_ty_utils/src/errors.rs` into `rustc_ty_utils/src/diagnostics.rs`
